+++ src/ServiceBus/ServiceBus/ChangeLog.md
@@ -19,6 +19,12 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added cmdlets for manual approval of Service Bus Private Endpoint Connections. The added cmdlets include,
+    -Approve-AzServiceBusPrivateEndpointConnection
+    -Deny-AzServiceBusPrivateEndpointConnection
+    -Get-AzServiceBusPrivateEndpointConnection
+    -Remove-AzServiceBusPrivateEndpointConnection
+    -Get-AzServiceBusPrivateLink
 
 ## Version 1.9.0
 * Fixed miscellaneous network rule set typos across module.